refactor: Move entity macro into its own file
